% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_class.R
\docType{class}
\name{AnyMatrix-class}
\alias{AnyMatrix-class}
\title{The OrganChat class}
\description{
The OrganChat object is created using the following inputs: (1) single-cell transcriptomics data matrix (normalized);
(2) organ names; (3) cell cluster and condition labels (named lists with cell barcodes).
}
