% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{CA_cal_scMeta.bygroup}
\alias{CA_cal_scMeta.bygroup}
\title{Calculate the mean expression level from single-cell metabolite data (for two conditions)}
\usage{
CA_cal_scMeta.bygroup(
  object,
  scMeta.data,
  organ.name = NULL,
  idents = NULL,
  conditions.label = NULL,
  mean_method = NULL,
  type = "amount",
  scale.max = 10
)
}
\arguments{
\item{object}{an OrganChat object}

\item{scMeta.data}{normalized metabolite amount matrix (metabolite-by-cell).}

\item{organ.name}{a char contains the organ names to be calculated.}

\item{idents}{a char of cluster labels to group the cells.}

\item{conditions.label}{a char of condition labels.}

\item{mean_method}{it is NULL by default and a weighted mean will be calculated (which is statisticallt more robust against noise).
If the input is "mean", then the arithmetic mean will be calculated.}

\item{type}{for metabolite data, OrganChat accepts two types of measurements, "amount" and "flux". The default type is "amount",
which means the data should be non-negative values. The "flux" data may have both negative (uptake) and positive (release) values.}

\item{scale.max}{The maximum value used to scale the data, which is 10 by default. No scaling will be performed if the input is NULL.}
}
\value{
an OrganChat object
}
\description{
If single-cell metabolite data is available, then this function can be applied to calculate the
average metabolite amount in each cell cluster in both conditions. The input must be the normalized amount matrix.
It should be applied to each organ if the data is available for multiple organs.
}
