% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{CA_cal_scRNA.bygroup}
\alias{CA_cal_scRNA.bygroup}
\title{Calculate the mean expression level from single-cell transcriptomics data (for two conditions)}
\usage{
CA_cal_scRNA.bygroup(
  object,
  DB = OrganChatDB,
  mean_method = NULL,
  scale.max = 10
)
}
\arguments{
\item{object}{an OrganChat object}

\item{DB}{the corresponding OrganChatDB used in the analysis.}

\item{mean_method}{it is NULL by default and a weighted mean will be calculated (which is statisticallt more robust against noise).
If the input is "mean", then the arithmetic mean will be calculated.}

\item{scale.max}{the maximum value used to scale the data, which is 10 by default. No scaling will be performed if the input is NULL.}
}
\value{
an OrganChat object
}
\description{
Single-cell transcriptomics data of two organs are requested to perform OOC. This function can be applied to
calculate the average gene expression of the long-range signals (LS) and downstream genes (DG) in each cell cluster, respectively (for both organs under BOTH conditions).
The LS and DG will be inferred using the corresponding OrganChatDB.
The outputs will be stored in "Object - LSdata/DGdata - organ".
}
