% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{CA_merge_metabolite.bygroup}
\alias{CA_merge_metabolite.bygroup}
\title{Merge existing cluster-level metabolite amount/flux data (for two conditions)}
\usage{
CA_merge_metabolite.bygroup(
  object,
  metabolitedata.list.condition1 = NULL,
  metabolitedata.list.condition2 = NULL,
  type = "amount",
  scale.max = 10
)
}
\arguments{
\item{object}{an OrganChat object}

\item{metabolitedata.list.condition1}{a list of metabolite data (average amount/flux for each cluster) for two organs in condition 1.}

\item{metabolitedata.list.condition2}{a list of metabolite data (average amount/flux for each cluster) for two organs in condition 2.}

\item{type}{for metabolite data, OrganChat accepts two types of measurements, "amount" and "flux". The default type is "amount",
which means the data should be non-negative values. The "flux" data may have both negative (uptake) and positive (release) values.}

\item{scale.max}{the maximum value used to scale the data, which is 10 by default. No scaling will be performed if the input is NULL.}
}
\value{
an OrganChat object
}
\description{
If the metabolite amount/flux data for both organs are ready, and are not calculated using th cal_scMeta.bygroup function,
they can still be integrated into the OrganChat object using this function for comparison analysis of two conditions.
This is usually because when single-cell metabolite data is unavailable, then the metabolite flux data can be inferred via METAFlux using OrganChat built-in modules.
Note the metabolites are only considered as the LS, the outputs will be stored in "Object - LSdata - organ".
}
