% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{CA_net0_infer}
\alias{CA_net0_infer}
\title{LS-R-SE-T pathway inference (for two conditions)}
\usage{
CA_net0_infer(
  object,
  DB = OrganChatDB,
  LS = NULL,
  receptor = NULL,
  tf = NULL,
  target = NULL
)
}
\arguments{
\item{object}{an OrganChat object}

\item{DB}{the corresponding OrganChatDB used in the analysis.}

\item{LS}{the list of LS used for inferring the pathways, the input is NULL by default, which means all will be used}

\item{receptor}{the list of Receptor used for inferring the pathways, the input is NULL by default, which means all will be used}

\item{tf}{the list of Signaling Effector used for inferring the pathways, the input is NULL by default, which means all will be used}

\item{target}{the list of Target used for inferring the pathways, the input is NULL by default, which means all will be used}
}
\value{
an OrganChat object
}
\description{
Infer LS-R-SE-T pathways using OrganChatDB, pathways with duplicated genes will be removed.
The output will be a data frame and stored in "Object - net0".
}
