% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{CA_net_analysis}
\alias{CA_net_analysis}
\title{Comparison analysis of two conditions}
\usage{
CA_net_analysis(
  object,
  K = 0.5,
  N = 2,
  R = 5,
  CI = 0.95,
  CA_cutoff_value = NULL,
  CA_cutoff_chatP = NULL,
  CA_cutoff_sd = NULL,
  CA_cutoff_CI = 0
)
}
\arguments{
\item{object}{an OrganChat object}

\item{K}{half-saturation constant of the Hill function, the default value is 0.5.}

\item{N}{Hill coefficient, the default value is 2.}

\item{R}{the number of bootstrap testing to perfomr, the default value is 5.}

\item{CI}{value of confidence interval, CI = 0.95 by default.}

\item{CA_cutoff_value}{a vector of 4 cutoff values for LS, R, SE, T, respectively. For each LS-T pathway, each component must have a higher
expression than the cutoff value in at least one condition. The default value is NULL.}

\item{CA_cutoff_chatP}{LS-T pathways with a communication probability lower than this cutoff value under both conditions will be removed.
The default value is NULL.}

\item{CA_cutoff_sd}{LS-T pathways with a standard deviation (calculated based on bootstrap results) higher than this cutoff value under at least one condition will be removed.
The default value is NULL.}

\item{CA_cutoff_CI}{S-T pathways with a total number of components lying within the confidence interval smaller than this value under at least one condition will be removed.
The default value is 0. The input must be 0,1,2, or 3.}
}
\value{
an OrganChat object
}
\description{
Calculate the communication probability, perform bootstrap tests, and filter the LS-T pathways for each organ and condition.
The output will be a data frame and stored in "Object - net".
}
