% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{cal_meanExpr}
\alias{cal_meanExpr}
\title{Calculate the mean expression level from a given expression matrix}
\usage{
cal_meanExpr(
  data,
  celluse = NULL,
  geneuse = NULL,
  idents = NULL,
  mean_method = NULL
)
}
\arguments{
\item{data}{normalized expression matrix (gene-by-cell).}

\item{celluse}{a char contains cell barcodes to be used to select cells. It is NULL by default, meaning all the cells will be used.}

\item{geneuse}{a char contains gene symbols to be used to select genes. It is NULL by default, meaning all the genes will be calculated.}

\item{idents}{a char of cluster labels to group the cells.}

\item{mean_method}{it is NULL by default and a weighted mean will be calculated (which is statisticallt more robust against noise).
If the input is "mean", then the arithmetic mean will be calculated.}
}
\value{
a data frame
}
\description{
Calculate the mean expression level in each cluster based on the normalized expression matrix (gene-by-cell), for one condition.
The output is a data frame (gene-by-cluster).
}
