% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{cal_netchat_bootstrap}
\alias{cal_netchat_bootstrap}
\title{Perform the bootstrap test (all LS-T pathways in the object)}
\usage{
cal_netchat_bootstrap(
  object,
  R = 5,
  sd_cutoff = 0.05,
  p_cutoff = NULL,
  CI_cutoff = 0,
  single.condition = TRUE,
  CI = 0.95
)
}
\arguments{
\item{object}{an OrganChat object}

\item{R}{the number of bootstrap testing to perfomr, the default value is 5.}

\item{sd_cutoff}{LS-T pathways with a standard deviation (calculated based on bootstrap results) higher than this cutoff value will be removed. The default value is 0.05.}

\item{p_cutoff}{LS-T pathways with a communication probability lower than this cutoff value will be removed. The default value is NULL.}

\item{CI_cutoff}{LS-T pathways with a total number of components lying within the confidence interval smaller than this value will be removed. The default value is 0.
The input must be 0,1,2, or 3.}

\item{single.condition}{the input is TRUE by default.}

\item{CI}{the value of confidence interval, CI = 0.95 by default.}
}
\value{
an OrganChat object
}
\description{
Perform the bootstrap test and calculated the standard deviation for all LS-T pathways stored under "object - net",
and filter them based on selected cutoff values.
}
