% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{cal_netchat_prob}
\alias{cal_netchat_prob}
\title{Calculate the communication probability (all LS-T pathways in the object)}
\usage{
cal_netchat_prob(object, K = 0.5, N = 2, single.condition = TRUE)
}
\arguments{
\item{object}{an OrganChat object}

\item{K}{half-saturation constant of the Hill function, the default value is 0.5.}

\item{N}{Hill coefficient, the default value is 2.}

\item{single.condition}{the input is TRUE by default.}
}
\value{
an OrganChat object
}
\description{
Calculate the communication probability of all LS-T pathways in the object, the signaling probability between two components are
modeled using Hill function x^N/(x^N+K^N).
}
