% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{cal_singlechat_bootstrap}
\alias{cal_singlechat_bootstrap}
\title{Perform the bootstrap test ((individual LS-T pathway))}
\usage{
cal_singlechat_bootstrap(object, R = 5, dir = NULL, CI = 0.95)
}
\arguments{
\item{object}{an OrganChat object}

\item{R}{the number of bootstrap testing to perfomr, the default value is 5.}

\item{dir}{the direction of comunication to analyze, The input should be one of the following: 'forward', 'backward', 'self_organ1', or 'self_organ2'.}

\item{CI}{value of confidence interval, CI = 0.95 by default.}
}
\value{
a list
}
\description{
Perform the bootstrap test and calculated the standard deviation for an individual LS-T pathway stored under "object - net".
}
