% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{cal_singlechat_prob}
\alias{cal_singlechat_prob}
\title{Calculate the communication probability (individual LS-T pathway)}
\usage{
cal_singlechat_prob(object, df, K = 0.5, N = 2, dir = NULL)
}
\arguments{
\item{object}{an OrganChat object}

\item{df}{a data frame of LS-T pathways.}

\item{K}{half-saturation constant of the Hill function, the default value is 0.5.}

\item{N}{Hill coefficient, the default value is 2.}

\item{dir}{the direction of comunication to analyze, The input should be one of the following: 'forward', 'backward', 'self_organ1', or 'self_organ2'.}
}
\value{
a data frame
}
\description{
Calculate the communication probability of an individual LS-T pathway, the signaling probability between two components are
modeled using Hill function x^N/(x^N+K^N).
}
