% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAFlux_processing.R
\name{calculate_avg_exp_Updated}
\alias{calculate_avg_exp_Updated}
\title{Updated METAFlux calculate_avg_exp function}
\usage{
calculate_avg_exp_Updated(myseurat, myident, n_bootstrap, seed)
}
\arguments{
\item{myseurat}{Seurat object. METAFlux will calculate on "data" slot}

\item{myident}{Seurat idents for grouping.This will be a character string indicating the grouping of the seurat object}

\item{n_bootstrap}{number of bootstrap}

\item{seed}{random seed}
}
\value{
mean expression data
}
\description{
Updated METAFlux calculate_avg_exp function to be compatible with Seurat V5
calculate_avg_exp: Calculate bootstrap mean expression for single cell data
}
