% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis_visualization.R
\name{CAOC_bubbleheatmap}
\alias{CAOC_bubbleheatmap}
\title{Bubble-heatmap showing two metrics of the cluster-level OOC of two conditions}
\usage{
CAOC_bubbleheatmap(
  object,
  dir = NULL,
  chatP_cutoff = NULL,
  category = "All",
  size = "ave.weight",
  color = "total.chatP",
  ncol = 2
)
}
\arguments{
\item{object}{an OrganChat object}

\item{dir}{the direction of comunication to analyze, The input should be one of the following: 'forward', 'backward', 'self_organ1', or 'self_organ2'.}

\item{chatP_cutoff}{LS-T pathways with a communication probability lower than this cutoff value will be removed. The default value is NULL.}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{size}{select which metric to showcase using the bubble size, the input should be one of the following:
"ave.weight", "total.counts", "total.chatP", or "ave.sd". The default value is "ave.weight".}

\item{color}{select which metric to showcase using the bubble color, the input should be one of the following:
"ave.weight", "total.counts", "total.chatP", or "ave.sd". The default value is "total.chatP".}

\item{ncol}{argument of the wrap_plots function, the default value is 2, meaning placing two plots in 2 columns, respectively.}
}
\value{
a ggplot2 object
}
\description{
Bubble-heatmap showing two metrics by bubble size and color of the cluster-level OOC of two conditions (using ggplot function).
}
