% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis_visualization.R
\name{CAOC_chorddiagram}
\alias{CAOC_chorddiagram}
\title{Chord diagrams of the cluster-level communications of two conditions}
\usage{
CAOC_chorddiagram(
  object,
  dir = NULL,
  category = "All",
  style = NULL,
  chatP_cutoff = NULL,
  big.gap = 20,
  directional = 1,
  grid.col = NULL,
  direction.type = c("diffHeight", "arrows"),
  link.arr.type = "big.arrow",
  link.target.prop = TRUE,
  small.gap = 1,
  annotationTrack = "grid",
  scale = FALSE,
  title.font = 2,
  title.cex = 0.2,
  title.line = 0,
  title.adj = 0.5,
  label.cex = 0.75,
  circle.margin = 0.5
)
}
\arguments{
\item{object}{an CAOC object (OrganChat object of comparison analysis).}

\item{dir}{specify which direction to select. The default value is NULL.}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{style}{style select which metric to showcase using the edge width, the input should be one of the following:
"ave.weight", "total.counts", "total.chatP", "ave.sd", or NULL (default, meaning "total.counts" will be used).}

\item{chatP_cutoff}{LS-T pathways with a communication probability lower than this cutoff value will be removed. The default value is NULL.}

\item{big.gap}{argument of the chordDiagram function, the default value is 20.}

\item{directional}{argument of the chordDiagram function, the default value is 1.}

\item{grid.col}{argument of the chordDiagram function, the default value is NULL.}

\item{direction.type}{argument of the chordDiagram function, the default value is c("diffHeight", "arrows").}

\item{link.arr.type}{argument of the chordDiagram function, the default value is  "big.arrow".}

\item{link.target.prop}{argument of the chordDiagram function, the default value is TRUE.}

\item{small.gap}{argument of the chordDiagram function, the default value is 1.}

\item{annotationTrack}{argument of the chordDiagram function, the default value is "grid".}

\item{scale}{argument of the chordDiagram function, the default value is FALSE.}

\item{title.font}{argument of the circos.trackPlotRegion function, the default value is 2.}

\item{title.cex}{argument of the circos.trackPlotRegion function, the default value is 0.2.}

\item{title.line}{argument of the circos.trackPlotRegion function, the default value is 0.}

\item{title.adj}{argument of the circos.trackPlotRegion function, the default value is 0.5.}

\item{label.cex}{argument of the circos.trackPlotRegion function, the default value is 0.75.}

\item{circle.margin}{argument of the circos.par function, the default value is 0.5.}
}
\description{
Chord diagrams of the cluster-level communications of two conditions.
}
