% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis_visualization.R
\name{CAOC_DI_barplot}
\alias{CAOC_DI_barplot}
\title{Plot the differential index (DI) of LS-T pathways}
\usage{
CAOC_DI_barplot(df, n = 5, axis.text.size = 20)
}
\arguments{
\item{df}{a data frame of differentially expressed LS-T pathways "object - DEpath".}

\item{n}{number of pathways to show in each cluster-cluster pair, the default value is 5.}

\item{axis.text.size}{argument of the theme function, the default value is 20.}
}
\description{
Bar plot of the differential index (DI) of LS-T pathways in each cluster-cluster pair. "DI>0: up-regulated in the first condition."
}
