% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{components_rosediagram}
\alias{components_rosediagram}
\title{Rose diagram to show the signaling inflow/outflow}
\usage{
components_rosediagram(
  df,
  category = "All",
  n = 3,
  outflow.colors = c("#e41a1c", "#377eb8", "#4daf4a"),
  inflow.colors = c("#377eb8", "#4daf4a", "#984ea3"),
  start.pos = 0.25 * pi,
  rotate.angle = TRUE,
  expand = FALSE,
  inner.radius = 0.1,
  r.axis.inside = TRUE,
  legend.position = "left",
  grid.width = 0.6
)
}
\arguments{
\item{df}{a data frame of the LS-T pathways in the OrganChat object.}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{n}{the number of components to show in each layer (LS, R, SE, T), the default value is 3.}

\item{outflow.colors}{argument of the scale_fill_manual function, defining the colors for LS, R, and SE. The default value is c("#e41a1c", "#377eb8", "#4daf4a").}

\item{inflow.colors}{argument of the scale_fill_manual function, defining the colors for R, SE, and T. The default value is c("#377eb8", "#4daf4a", "#984ea3").}

\item{start.pos}{argument of the coord_radial function, the default value is 0.25*pi.}

\item{rotate.angle}{argument of the coord_radial function, the default value is TRUE.}

\item{expand}{argument of the coord_radial function, the default value is FALSE.}

\item{inner.radius}{argument of the coord_radial function, the default value is 0.1.}

\item{r.axis.inside}{argument of the coord_radial function, the default value is TRUE.}

\item{legend.position}{argument of the theme function, the default value is "left".}

\item{grid.width}{argument of the grid.arrange function, the default value is 0.6.}
}
\description{
Rose diagram to show the components of a cluster-cluster communication with the most inflow/outflow (using ggplot2).
}
