% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{constrcuct_EdgeNode_df}
\alias{constrcuct_EdgeNode_df}
\title{Construct a list of data frames recording the edges & nodes from a LS-T table}
\usage{
constrcuct_EdgeNode_df(df)
}
\arguments{
\item{df}{a data frame of the LS-T pathways in the OrganChat object}
}
\value{
a list
}
\description{
From any data frame of the LS-T pathways in the OrganChat object, construct a list of data frames recording the directed & weighted network.
The direction indicates the activation along the LS-R-SE-T path, the weight is predicted based on the Hill function model.
}
