% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{construct_EdgeNode}
\alias{construct_EdgeNode}
\title{Construct the edge and node data.frames}
\usage{
construct_EdgeNode(
  object,
  category = "All",
  chatP_cutoff = NULL,
  single.condition = TRUE
)
}
\arguments{
\item{object}{an OrganChat object}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{chatP_cutoff}{LS-T pathways with a communication probability lower than this cutoff value will be removed. The default value is NULL.}

\item{single.condition}{the input is TRUE by default.}
}
\value{
a list
}
\description{
Construct the edge and node data.frames from all data frames stored under "object - net".
}
