% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{construct_MTmaxflow_df}
\alias{construct_MTmaxflow_df}
\title{Calculate the signal inflow & outflow of a GRN}
\usage{
construct_MTmaxflow_df(df)
}
\arguments{
\item{df}{a data frame of the LS-T pathways in the OrganChat object}
}
\value{
a data frame
}
\description{
For a cluster-cluster communication table of one direction, calculate the following three metrics for each "LS & T" pair: (1) total counts -- total number of pathways involved;
(2) ave. sd -- average bootstrap sd of these pathways; (3) max flow --  maximum flow from the LS to the T (using the graph_from_data_frame function).
}
