% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{construct_Network_df}
\alias{construct_Network_df}
\title{Construct a list for the directed & weighted GRN with user-selected filters.}
\usage{
construct_Network_df(df, category = "All", restrict.by = "Both")
}
\arguments{
\item{df}{a data frame of the LS-T pathways in the OrganChat object}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{restrict.by}{use this parameter to indicate whether each GRN has a single LS, a single Target, or both. The input should be one of the following:
"LS", "Target", "Both". The default value is "Both"}
}
\value{
a list
}
\description{
This function is basically constrcuct_EdgeNode_df function with additional options, generating a list of outputs from constrcuct_EdgeNode_df function.
}
