% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_class.R
\name{create_OrganChat}
\alias{create_OrganChat}
\title{Create an OrganChat object}
\usage{
create_OrganChat(
  object,
  organs = NULL,
  idents.list = NULL,
  ComparisonAnalysis = FALSE,
  conditions = NULL,
  conditions.list = NULL,
  assay = NULL,
  do.sparse = T
)
}
\arguments{
\item{object}{a normalized (NOT count) data matrix (gene-by-cell)}

\item{organs}{a char contains organ names (should be the same names recorded in the meta data)}

\item{idents.list}{a list of cluster labels (named with cell barcodes) for each organ}

\item{ComparisonAnalysis}{a boolean value to determine if the comparison between condition is performing,
comparison analysis will be used for multi-organ (more than 2) scenario. Only the clusters shared by both conditions will be used for comparison analysis.}

\item{conditions}{a char contains condition labels (should be the same names recorded in the meta data)}

\item{conditions.list}{a list of condition labels (named with cell barcodes) for each organ}

\item{assay}{an assay to use when the input is a Seurat object}

\item{do.sparse}{whether use sparse format}
}
\value{
an OrganChat object
}
\description{
Create an OrganChat object using single-cell transcriptomics data matrix (normalized), meta data of organ, condition, and cluster.
}
