% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{DEpath_table}
\alias{DEpath_table}
\title{Infer differential expressed pathways}
\usage{
DEpath_table(
  object,
  Sender.group = NULL,
  Receiver.group = NULL,
  Direction = NULL,
  p_cutoff = NULL,
  sd_cutoff = NULL,
  RI_cutoff = NULL,
  top = NULL
)
}
\arguments{
\item{object}{an OrganChat object}

\item{Sender.group}{specify which Sender group to select. The default value is NULL.}

\item{Receiver.group}{specify which Receiver group to select. The default value is NULL.}

\item{Direction}{specify which direction to select. The default value is NULL.}

\item{p_cutoff}{LS-T pathways with a communication probability lower than this cutoff value under both conditions will be removed.
The default value is NULL.}

\item{sd_cutoff}{LS-T pathways with a standard deviation (calculated based on bootstrap results) higher than this cutoff value under at least one condition will be removed.
The default value is NULL.}

\item{RI_cutoff}{keep the LS-T pathways which the absolute RI value is higher then this cutoff value. The default value is NULL.}

\item{top}{keep the top pathways based on the positive and negative RI, respectively. The default value is NULL.}
}
\value{
an OrganChat object
}
\description{
Generate a data frame for the LS-T pathways for comparison between conditions
The output will be a data frame and stored in "Object - DEpath".
}
