% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{Export_path_table}
\alias{Export_path_table}
\title{Export pathways}
\usage{
Export_path_table(object)
}
\arguments{
\item{object}{an OrganChat object}
}
\value{
a data frame
}
\description{
Export all inferred pathways in one table
}
