% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{filter_OCnet0}
\alias{filter_OCnet0}
\title{LS-R-SE-T pathway filtering for all directions}
\usage{
filter_OCnet0(object, cutoff_value = NULL, single.condition = TRUE)
}
\arguments{
\item{object}{an OrganChat object}

\item{cutoff_value}{a vector of 4 cutoff values for LS, R, SE, T, respectively.}

\item{single.condition}{the input is TRUE by default.}
}
\value{
an OrganChat object
}
\description{
Filter the 'net0' by removing lowly/zero expressed LS & DG for all four directions, one condition.
The output will be stored under "object - net".
}
