% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{filter_OCnet0_onedir}
\alias{filter_OCnet0_onedir}
\title{LS-R-SE-T pathway filtering (for one direction)}
\usage{
filter_OCnet0_onedir(
  object,
  cutoff_value = NULL,
  dir = NULL,
  single.condition = TRUE
)
}
\arguments{
\item{object}{an OrganChat object}

\item{cutoff_value}{a vector of 4 cutoff values for LS, R, SE, T, respectively.}

\item{dir}{the direction of comunication to analyze, The input should be one of the following: 'forward', 'backward', 'self_organ1', or 'self_organ2'.}

\item{single.condition}{the input is TRUE by default.}
}
\value{
a list
}
\description{
Filter the 'net0' by removing lowly/zero expressed LS & DG for one direction, one condition.
The output will be a list of data frames for cluster-cluster communications.
}
