% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{filter_OCreceptor}
\alias{filter_OCreceptor}
\title{Remove the receptors activated by non-LS signals}
\usage{
filter_OCreceptor(
  object,
  DB = OrganChatDB,
  LRDB = LRDB,
  mean_method = NULL,
  exp_cutoff = NULL,
  single.condition = TRUE
)
}
\arguments{
\item{object}{an OrganChat object}

\item{DB}{the corresponding OrganChatDB used in the analysis.}

\item{LRDB}{an additional ligand-receptor database, such as CellChatDB, should be a data frame contains ligand-receptor pairs.}

\item{mean_method}{it is NULL by default and a weighted mean will be calculated (which is statisticallt more robust against noise).
If the input is "mean", then the arithmetic mean will be calculated.}

\item{exp_cutoff}{a scalar used to identify highly expressed ligands.}

\item{single.condition}{the input is TRUE by default.}
}
\value{
an OrganChat object
}
\description{
(Optional step) Remove the receptors (and the associated LS-T pathways) that: (1) they can be activated by the ligands outside
the OrganChatDB, based on other ligand-receptor sources such as CellChatDB; (2) these ligands are recorded in the data and highly expressed.
The additional ligand-receptor database is need (LRDB) to perform this step.
}
