% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAFlux_processing.R
\name{generate_boots}
\alias{generate_boots}
\title{Updated METAFlux generate_boots function}
\usage{
generate_boots(celltype, n)
}
\arguments{
\item{celltype}{colnames of single cell data.Colnames should be labeled as cell type or cluster.}

\item{n}{number of bootstrap}
}
\value{
a data frame
}
\description{
Updated METAFlux generate_boots function to be compatible with Seurat V5
generate_boots: Generate bootstrap index data
}
