% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAFlux_processing.R
\name{get_ave_exp_Updated}
\alias{get_ave_exp_Updated}
\title{Updated METAFlux get_ave_exp function}
\usage{
get_ave_exp_Updated(i, myseurat, samples, myident)
}
\arguments{
\item{i}{index}

\item{myseurat}{single cell Seurat object.METAFlux will calculate on "data" slot}

\item{samples}{generated bootstrap index data}

\item{myident}{Seurat idents.This will be a character string indicating the grouping of the seurat object}
}
\value{
a data frame
}
\description{
Updated METAFlux get_ave_exp function to be compatible with Seurat V5
get_ave_exp: Calculate mean expression for one bootstrap
}
