% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{GOanalysis_scatterplot}
\alias{GOanalysis_scatterplot}
\title{Scatter plot of the top associated GO terms}
\usage{
GOanalysis_scatterplot(df, n = 10, size = 6)
}
\arguments{
\item{df}{a data frame of the OC_GOanalysis output.}

\item{n}{total number of GO terms to show, the default value is 10.}

\item{size}{size of the dots, the default value is 6.}
}
\description{
Scatter plot via ggplot2 to show the top-related GO terms based on the OC_GOanalysis output.
}
