% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{merge_metabolite.bygroup}
\alias{merge_metabolite.bygroup}
\title{Merge existing cluster-level metabolite amount/flux data}
\usage{
merge_metabolite.bygroup(
  object,
  metabolitedata.list = NULL,
  scale.max = 10,
  type = "flux",
  single.condition = TRUE
)
}
\arguments{
\item{object}{an OrganChat object}

\item{metabolitedata.list}{a list of metabolite data (average amount/flux for each cluster) for two organs}

\item{scale.max}{the maximum value used to scale the data, which is 10 by default. No scaling will be performed if the input is NULL.}

\item{type}{for metabolite data, OrganChat accepts two types of measurements, "amount" and "flux". The default type is "amount",
which means the data should be non-negative values. The "flux" data may have both negative (uptake) and positive (release) values.}

\item{single.condition}{the input is TRUE by default.}
}
\value{
an OrganChat object
}
\description{
If the metabolite amount/flux data for both organs are ready, and are not calculated using th cal_scMeta.bygroup function,
they can still be integrated into the OrganChat object using this function.
This is usually because when single-cell metabolite data is unavailable, then the metabolite flux data can be inferred via METAFlux using OrganChat built-in modules.
Note the metabolites are only considered as the LS, the outputs will be stored in "Object - LSdata - organ".
}
