% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAFlux_processing.R
\name{METAFlux_convert_OrganChat}
\alias{METAFlux_convert_OrganChat}
\title{Convert the processed METAFlux output for OrganChat analysis}
\usage{
METAFlux_convert_OrganChat(
  cell.types,
  flux,
  METAFlux_lookup = METAFlux_lookup,
  hmdb_dictionary = OrganChatDB_hmdb_dictionary
)
}
\arguments{
\item{cell.types}{a char of cell cluster levels}

\item{flux}{the processed and reformated METAFlux output table}

\item{METAFlux_lookup}{use the METAFlux_lookup database to match the reaction ID.}

\item{hmdb_dictionary}{OrganChat-provided database matching "metabolite name" and the HMDB ID.}
}
\value{
a data frame
}
\description{
Match the ReactionID (Metabolite name) to the HMDB ID, construct the data frame (row--metabolites (LS), col--clusters) that is ready
to be stored under "OrganChat object - LSdata - organ".
}
