% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAFlux_processing.R
\name{METAFlux_reformat}
\alias{METAFlux_reformat}
\title{Process and reformat the METAFlux output}
\usage{
METAFlux_reformat(
  flux,
  cell.types,
  exclude = "None",
  METAFlux_lookup = METAFlux_lookup,
  scale.byflux = TRUE
)
}
\arguments{
\item{flux}{the METAFlux output table}

\item{cell.types}{a char of cell cluster levels}

\item{exclude}{specify if neglect the the reactions driven by external_medium, internal_medium, or both.
The input should be one of the following: "None", "external_medium", "internal_medium", "Both". The default value is "None".}

\item{METAFlux_lookup}{use the METAFlux_lookup database to match the reaction ID.}

\item{scale.byflux}{a boolean value to determine if scale the average influx and outflux data to 0 and 1 or -1 and 0, respectively.}
}
\value{
a data frame
}
\description{
Calculate the average flux value based on the bootstrap results; scale the influx and outflux, respectively; adding cluster labels, reaction IDs, etc.
}
