% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOOC_visualization.R
\name{MO_networkconstruct}
\alias{MO_networkconstruct}
\title{Construct a list for the directed & weighted GRN from the output of MO_pairwise_analysis function.}
\usage{
MO_networkconstruct(object, MO.net.list, category = "All", chatP_cutoff = NULL)
}
\arguments{
\item{object}{an OrganChat object, normally the output of MO_merge_metabolite.bygroup function.}

\item{MO.net.list}{output (a list) of the MO_pairwise_analysis function.}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{chatP_cutoff}{LS-T pathways with a communication probability lower than this cutoff value will be removed. The default value is NULL.}
}
\value{
a list
}
\description{
This function generates a list of directed & weighted GRNs from MO_pairwise_analysis function outputs. Each will have a data frame for edges and nodes, respectively.
}
