% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOOC_visualization.R
\name{MO_networkplot}
\alias{MO_networkplot}
\title{Chord diagrams of the organ-level communications of two conditions}
\usage{
MO_networkplot(
  network.input,
  style = "total.counts",
  color.platte = brewer.pal(8, "Set2"),
  scale.edge.width = 5,
  scale.vertex.size = 30,
  vertex.frame.color = "white",
  vertex.label.color = "black",
  vertex.label.dist = 3,
  edge.arrow.size = 0.6,
  edge.curved = 0.2,
  margin = 0.1,
  vertex.max = NULL,
  edge.max = NULL
)
}
\arguments{
\item{network.input}{output (a list) of the MO_networkconstruct function.}

\item{style}{style select which metric to showcase using the edge width, the input should be one of the following:
"ave.weight", "total.counts", "total.chatP", "ave.sd", or NULL (default, meaning "total.counts" will be used).}

\item{color.platte}{= brewer.pal(8, "Set2"),}

\item{scale.edge.width}{a multiplier to scale the edge width, the default value is 5.}

\item{scale.vertex.size}{a multiplier to scale the vertex size, the default value is 30.}

\item{vertex.frame.color}{argument of the plot function, the default value is "white".}

\item{vertex.label.color}{argument of the plot function, the default value is "black".}

\item{vertex.label.dist}{argument of the plot function, the default value is 3.}

\item{edge.arrow.size}{argument of the plot function, the default value is 0.6.}

\item{edge.curved}{argument of the plot function, the default value is 0.2.}

\item{margin}{argument of the plot function, the default value is 0.1.}

\item{vertex.max}{scale the vertex size using the maximum in "condition1" or "condition2" or NULL (default, meaning using the maximum of both conditions).}

\item{edge.max}{scale the edge width using the maximum in "condition1" or "condition2" or NULL (default, meaning using the maximum of both conditions).}
}
\description{
Chord diagrams of the organ-level communications of two conditions.
}
