% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{network_treeplot}
\alias{network_treeplot}
\title{Plot the network structure}
\usage{
network_treeplot(
  df,
  edge.width.scalemax = 2,
  node.size.scalemax = 20,
  edge.color = "grey50",
  edge.arrow.size = 0.5,
  edge.arrow.width = 0.5,
  vertex.frame.color = "white",
  vertex.label.color = "black",
  vertex.label.cex = 0.65,
  vertex.label.dist = 0,
  vertex.label.degree = 0
)
}
\arguments{
\item{df}{a data frame of the LS-T pathways in the OrganChat object.}

\item{edge.width.scalemax}{scale the edge width to a selected maximum value, the default value is 2.}

\item{node.size.scalemax}{scale the node size to a selected maximum value, the default value is 20.}

\item{edge.color}{argument of the plot function, the default value is "grey50".}

\item{edge.arrow.size}{argument of the plot function, the default value is 0.5.}

\item{edge.arrow.width}{argument of the plot function, the default value is 0.5.}

\item{vertex.frame.color}{argument of the plot function, the default value is "white".}

\item{vertex.label.color}{argument of the plot function, the default value is "black".}

\item{vertex.label.cex}{argument of the plot function, the default value is 0.65.}

\item{vertex.label.dist}{argument of the plot function, the default value is 0.}

\item{vertex.label.degree}{argument of the plot function, the default value is 0.}
}
\description{
Plot the network structure (using igraph package), the input should be one GRN of the output from 'construct_Network_df' function.
}
