% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_analysis.R
\name{OC_GOanalysis}
\alias{OC_GOanalysis}
\title{GO enrichment analysis}
\usage{
OC_GOanalysis(
  df,
  LS = NULL,
  layer = NULL,
  organism = "mmusculus",
  ordered_query = TRUE,
  multi_query = FALSE,
  significant = TRUE,
  exclude_iea = FALSE,
  measure_underrepresentation = FALSE,
  evcodes = FALSE,
  user_threshold = 0.05,
  correction_method = "g_SCS",
  domain_scope = "annotated",
  custom_bg = NULL,
  numeric_ns = "",
  sources = NULL,
  as_short_link = FALSE,
  highlight = FALSE
)
}
\arguments{
\item{df}{a data frame of LS-T pathways}

\item{LS}{selected LS to be included in the GO analysis. The default value is NULL, meaning all will be used.}

\item{layer}{specify which of these will be included: "Receptor", "TF", or "Target". The default value is NULL, meaning all will be used.}

\item{organism}{the species used in the analysis. The default value is "mmusculus".}

\item{ordered_query}{argument of the gost function in gprofiler2, the default value is TRUE.}

\item{multi_query}{argument of the gost function in gprofiler2, the default value is FALSE.}

\item{significant}{argument of the gost function in gprofiler2, the default value is TRUE.}

\item{exclude_iea}{argument of the gost function in gprofiler2, the default value is FALSE.}

\item{measure_underrepresentation}{argument of the gost function in gprofiler2, the default value is FALSE.}

\item{evcodes}{argument of the gost function in gprofiler2, the default value is FALSE.}

\item{user_threshold}{argument of the gost function in gprofiler2, the default value is 0.05.}

\item{correction_method}{argument of the gost function in gprofiler2, the default value is "g_SCS".}

\item{domain_scope}{argument of the gost function in gprofiler2, the default value is "annotated".}

\item{custom_bg}{argument of the gost function in gprofiler2, the default value is NULL.}

\item{numeric_ns}{argument of the gost function in gprofiler2, the default value is "".}

\item{sources}{argument of the gost function in gprofiler2, the default value is NULL.}

\item{as_short_link}{argument of the gost function in gprofiler2, the default value is FALSE.}

\item{highlight}{argument of the gost function in gprofiler2, the default value is FALSE.}
}
\value{
a list
}
\description{
Perform GO enrichment analysis of a data frame of LS-T pathways using the package "gprofiler2".
}
