% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{OC_riverplot}
\alias{OC_riverplot}
\title{River plot of selected LS-T pathways}
\usage{
OC_riverplot(
  df,
  pref = "LS",
  category = "All",
  n.path = 20,
  text.color = TRUE,
  LS_name = TRUE,
  width = 1/3,
  size = 0.1,
  expand = c(0.01, 0.05),
  type = "viridis",
  palette = "RdPu"
)
}
\arguments{
\item{df}{a data frame of LS-T pathways from OrganChat object}

\item{pref}{a char indicating the preference when selecting the pathways, the input should be one of the following: "chatP", "LS", "Receptor", "TF", "Target"
The default value is "LS", meaning OrganChat will try to cover as many LS as possible, by selecting one LS-T pathway of each LS (ranked by the communication probability) until reaching the limit n.path.
If the input is "chatP", then OrganChat will select the top n.path LS-T pathways based on the communication probability.}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{n.path}{total number pathways to plot, the default value is 20.}

\item{text.color}{the default value is TRUE.}

\item{LS_name}{a boolean value to specify if the metabolite name is used, the default value is TRUE. Otherwise the HMDB ID will be presented.}

\item{width}{argument of the geom_stratum function, the default value is 1/3.}

\item{size}{argument of the geom_stratum function, the default value is 0.1.}

\item{expand}{argument of the scale_x_discrete function, the default value is c(.01, .05).}

\item{type}{argument of the scale_fill_continuous function, the default value is "viridis".}

\item{palette}{argument of the scale_color_distiller function, the default value is "RdPu".}
}
\value{
a ggplot2 object
}
\description{
River plot of the selected pathways from a data.frame (using ggplot), based on user-select preference and LS type.
}
