% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{OC_singledir_netanalysis}
\alias{OC_singledir_netanalysis}
\title{Select LS-T pathways from one direction table for plot}
\usage{
OC_singledir_netanalysis(
  object,
  net,
  dir,
  chatP_cutoff = NULL,
  category = "All"
)
}
\arguments{
\item{object}{an OrganChat object}

\item{net}{a list of data frames under "object - net - Condition - direction" layer}

\item{dir}{the direction of comunication to analyze, The input should be one of the following: 'forward', 'backward', 'self_organ1', or 'self_organ2'.}

\item{chatP_cutoff}{LS-T pathways with a communication probability lower than this cutoff value will be removed. The default value is NULL.}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}
}
\value{
a data frame
}
\description{
Select LS-T pathways from one direction, calculate the total count (total.counts), average communication probability (ave.weight), and total communication probability (total.chatP).
}
