% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganChat_class.R
\docType{class}
\name{OrganChat-class}
\alias{OrganChat-class}
\alias{object}
\title{The key slots used in the OrganChat object are described below.}
\description{
The key slots used in the OrganChat object are described below.
}
\section{Slots}{

\describe{
\item{\code{organs}}{a char contains organ names (should be the same names recorded in the meta data)}

\item{\code{conditions}}{a char contains condition labels (should be the same names recorded in the meta data)}

\item{\code{data.list}}{a list of single-cell data matrices for each organ}

\item{\code{idents.list}}{a list of cluster labels (named with cell barcodes) for each organ}

\item{\code{conditions.list}}{a list of condition labels (named with cell barcodes) for each organ.
The input 'conditions.list' must have the same levels listed in the input "conditions".}

\item{\code{LSdata}}{a list of data frame storing the cluster-level average expression of the long-range signal (LS) for each organ}

\item{\code{DGdata}}{a list of data frame storing the cluster-level average expression of the downstream gene (DG, including R, SE, T) for each organ}

\item{\code{net0}}{a data frame storing all possible LS-T pathways inferred based on the OrganChatDB, without any filtering}

\item{\code{net}}{a list of data frame with multiple layers, storing the filtered LS-T pathways for each cluster-cluster pair, in each condition and organ}

\item{\code{DEpath}}{a data frame storing the comparison results of the LS-T pathways between conditions}

\item{\code{options}}{a list of miscellaneous data, such as parameters used throughout analysis
#'}
}}

