% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisonanalysis.R
\name{quant_vector}
\alias{quant_vector}
\title{Calculate the quantile position}
\usage{
quant_vector(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
Calculate the quantile position for every number in vector x
}
