% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAFlux_processing.R
\name{run_METAFlux}
\alias{run_METAFlux}
\title{Perform METAFlux analysis in OrganChat}
\usage{
run_METAFlux(
  Data.input,
  myident,
  n_bootstrap = 5,
  seed = 1,
  medium = human_blood
)
}
\arguments{
\item{Data.input}{Seurat object}

\item{myident}{Seurat idents for grouping.This will be a character string indicating the grouping of the seurat object}

\item{n_bootstrap}{number of the bootstrap test, the default value is 5.}

\item{seed}{random seed, the default value is 1.}

\item{medium}{= human_blood}
}
\value{
mean expression data
}
\description{
Run METAFlux in OrganChat with updated functions
}
