% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCanalysis_visualization.R
\name{select_paths}
\alias{select_paths}
\title{Select LS-T pathways for visualization}
\usage{
select_paths(df, pref = "LS", category = "All", n.path = 20)
}
\arguments{
\item{df}{a data frame of LS-T pathways}

\item{pref}{a char indicating the preference when selecting the pathways, the input should be one of the following: "chatP", "LS", "Receptor", "TF", "Target"
The default value is "LS", meaning OrganChat will try to cover as many LS as possible, by selecting one LS-T pathway of each LS (ranked by the communication probability) until reaching the limit n.path.
If the input is "chatP", then OrganChat will select the top n.path LS-T pathways based on the communication probability.}

\item{category}{use this parameter to select the types of LS. The input should be one of the following: "All", "hormone_peptide", "cytokines", "hormone_nonpeptide", "metabolite".
The default value is "All", meaning all the LS will be included.}

\item{n.path}{total number pathways to plot, the default value is 20.}
}
\value{
a data frame
}
\description{
Select the preferred pathways from a data.frame, based on user-select preference and LS type.
}
